# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/challenge": {
      "post": {
        "operationId": "AuthController_getChallenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Get challenge message for wallet signature",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifySignatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify signature and issue JWT tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Logout and revoke refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout from all sessions",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user information",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/email/add": {
      "post": {
        "operationId": "AuthController_addEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add an email address and send verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/email/verify": {
      "post": {
        "operationId": "AuthController_verifyEmailAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify an email with a 6-digit code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/email/resend": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resend email verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/email/{id}": {
      "delete": {
        "operationId": "AuthController_removeEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an email address",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/email/{id}/primary": {
      "put": {
        "operationId": "AuthController_setPrimaryEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set an email as primary",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/payments/verify": {
      "post": {
        "operationId": "PaymentsController_verifyPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify a USDC payment on Base",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/paystack/initialize": {
      "post": {
        "operationId": "PaymentsController_initializePaystack",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializePaystackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initialize a Paystack payment",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/paystack/verify": {
      "post": {
        "operationId": "PaymentsController_verifyPaystack",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify a Paystack payment",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/flutterwave/initialize": {
      "post": {
        "operationId": "PaymentsController_initializeFlutterwave",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initialize a Flutterwave payment",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/flutterwave/verify": {
      "post": {
        "operationId": "PaymentsController_verifyFlutterwave",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify a Flutterwave payment",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/flutterwave/webhook": {
      "post": {
        "operationId": "PaymentsController_handleFlutterwaveWebhook",
        "parameters": [
          {
            "name": "verf-hash",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "verif-hash",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Handle Flutterwave webhooks",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/chainrails/quote": {
      "post": {
        "operationId": "PaymentsController_getChainrailsQuotes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChainrailsQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cross-chain payment quotes from Chainrails",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/chainrails/create-intent": {
      "post": {
        "operationId": "PaymentsController_createChainrailsIntent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChainrailsIntentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Chainrails cross-chain payment intent",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/chainrails/webhook": {
      "post": {
        "operationId": "PaymentsController_handleChainrailsWebhook",
        "parameters": [
          {
            "name": "x-chainrails-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-chainrails-timestamp",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Handle Chainrails webhooks",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/v1/admin/auth/me": {
      "get": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin profile",
        "tags": [
          "Admin Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "EigenWatch API",
    "description": "API for comprehensive EigenLayer operator analytics, performance tracking, and network insights",
    "version": "1.0",
    "contact": {
      "name": "EigenWatch Team",
      "url": "https://eigenwatch.xyz",
      "email": "eigenwatchteam@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development"
    },
    {
      "url": "https://api-staging.eigenwatch.xyz",
      "description": "Staging"
    },
    {
      "url": "https://api.eigenwatch.xyz",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      },
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT token from wallet signature"
      }
    },
    "schemas": {
      "ChallengeRequestDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Ethereum wallet address",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          }
        },
        "required": [
          "address"
        ]
      },
      "VerifySignatureDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Ethereum wallet address",
            "example": "0x1234567890abcdef1234567890abcdef12345678"
          },
          "signature": {
            "type": "string",
            "description": "Signed message signature"
          },
          "nonce": {
            "type": "string",
            "description": "Nonce from challenge response"
          }
        },
        "required": [
          "address",
          "signature",
          "nonce"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token"
          }
        }
      },
      "AddEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address to add",
            "example": "user@example.com"
          },
          "risk_alerts": {
            "type": "boolean",
            "description": "Opt-in for risk alerts",
            "default": true
          },
          "marketing": {
            "type": "boolean",
            "description": "Opt-in for marketing emails",
            "default": false
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address to verify",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "description": "6-digit verification code",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address to resend verification for",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "VerifyPaymentDto": {
        "type": "object",
        "properties": {
          "txHash": {
            "type": "string",
            "description": "The transaction hash of the USDC transfer on Base",
            "example": "0x..."
          }
        },
        "required": [
          "txHash"
        ]
      },
      "InitializePaystackDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email address of the user making the payment"
          }
        },
        "required": [
          "email"
        ]
      },
      "ChainrailsQuoteDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "description": "The amount to transfer",
            "example": "20"
          },
          "destinationChain": {
            "type": "string",
            "description": "The destination chain identifier",
            "example": "BASE_MAINNET"
          },
          "tokenOut": {
            "type": "string",
            "description": "The token contract address on the destination chain",
            "example": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
          }
        },
        "required": [
          "amount",
          "destinationChain",
          "tokenOut"
        ]
      },
      "ChainrailsIntentDto": {
        "type": "object",
        "properties": {
          "sender": {
            "type": "string",
            "description": "The sender wallet address",
            "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
          },
          "amount": {
            "type": "string",
            "description": "The amount to transfer (in smallest unit or human-readable)",
            "example": "20000000"
          },
          "amountSymbol": {
            "type": "string",
            "description": "The symbol denomination of the amount (e.g. USDC). Defaults to USDC if omitted.",
            "example": "USDC"
          },
          "tokenIn": {
            "type": "string",
            "description": "The input token contract address on the source chain",
            "example": "0x036CbD53842c5426634e7929541eC2318f3dCF7e"
          },
          "sourceChain": {
            "type": "string",
            "description": "The source chain identifier",
            "example": "ARBITRUM_MAINNET"
          },
          "destinationChain": {
            "type": "string",
            "description": "The destination chain identifier",
            "example": "BASE_MAINNET"
          },
          "recipient": {
            "type": "string",
            "description": "The recipient wallet address on the destination chain",
            "example": "0xb79541Be080a59fdcE6C0b43219ba56c725eC65e"
          },
          "refundAddress": {
            "type": "string",
            "description": "The refund address if the intent fails or expires",
            "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata to attach to the intent"
          }
        },
        "required": [
          "sender",
          "amount",
          "tokenIn",
          "sourceChain",
          "destinationChain",
          "recipient",
          "refundAddress"
        ]
      },
...
    }
  }
}